Feature: FN Kernel Using LibAsync
In this commit, we introduce a new version for the FN kernel, which
relies on the LibAsync that is shipped with MPPA-256. Overall, instead
of relying on synchronous communication primitives, this version of the
kernel relies on asynchronous communications and thereby tries to
overlap as much as possible computation with communication.